utils - separate for node and common
