+++ sdk/test-utils/recorder/src/utils.node.ts
@@ -1,0 +1,58 @@
+import fs from "fs-extra";
+import path from "path";
+
+/**
+ * ONLY WORKS IN THE NODE.JS ENVIRONMENT
+ *
+ * Meant to be called during the playback for the node tests.
+ * 1. Takes the test filePath as argument.
+ * 2. Looks for the `recordings` folder in its hierarchical path.
+ * 3. Returns the full path of the `recordings` folder
+ *
+ * While running the tests, `filePath` can vary depending on location of the test files, examples below
+ *
+ * 1. If roll-up generated bundle files are being leveraged to run the tests
+ *    filePath = `<base path>\azure-sdk-for-js\sdk\storage\storage-blob\dist-test\index.node.js`
+ * 2. If ts complied dist-esm files are being used to run the tests
+ *    filePath = `<base path>\azure-sdk-for-js\sdk\storage\storage-blob\dist-esm\test\utils.spec.js`
+ *    filePath = `<base path>\azure-sdk-for-js\sdk\storage\storage-blob\dist-esm\test\node\utils.spec.js`
+ * 3. If `.spec.ts` test files are being used directly
+ *    filePath = `<base path>\azure-sdk-for-js\sdk\storage\storage-blob\test\utils.spec.ts`
+ *    filePath = `<base path>\azure-sdk-for-js\sdk\storage\storage-blob\test\node\utils.spec.ts`
+ * In the above example, no matter where the test files are,
+ *    the recordings are located at `<base path>\azure-sdk-for-js\sdk\storage\storage-blob\recordings\`.
+ * In order to playback the tests, exact location of the recordings is to be found,
+ *    this is done by checking the parent(s) folders until the `recordings` folder is found.
+ *
+ * @export
+ * @param {string} filePath
+ * @returns {string} location of the `recordings` folder
+ */
+export function findRecordingsFolderPath(filePath: string): string {
+  // Stripping away the file name
+  let currentPath = path.resolve(filePath, "..");
+  // File/folder path of a closest child of `currentPath` in the folder hierarchy of `filePath`
+  let lastPath = filePath;
+  try {
+    // While loop to find the `recordings` folder
+    while (!fs.existsSync(path.resolve(currentPath, "recordings/"))) {
+      if (fs.existsSync(path.resolve(currentPath, "package.json"))) {
+        // package.json of the SDK is found but not the `recordings` folder
+        // which is supposed to be present at the same level as package.json
+        throw new Error(`'recordings' folder is not found at ${currentPath}`);
+      } else if (lastPath === currentPath) {
+        throw new Error(
+          `'recordings' folder is not found at ${currentPath} (reached the root directory)`
+        );
+      } else {
+        lastPath = currentPath;
+        currentPath = path.resolve(currentPath, "..");
+      }
+    }
+    return path.resolve(currentPath, "recordings/");
+  } catch (error) {
+    throw new Error(
+      `Unable to locate the 'recordings' folder anywhere in the hierarchy of the file path ${filePath}\n ${error}`
+    );
+  }
+}
